+++ src/strategies/bear_put_spread.rs
@@ -26,6 +26,7 @@ use super::base::{
     BreakEvenable, Optimizable, Positionable, Strategable, StrategyBasics, StrategyType, Validable,
 };
 use super::shared::SpreadStrategy;
+use crate::strategies::base::lower_break_even;
 use crate::{
     ExpirationDate, Options,
     chains::{StrategyLegs, chain::OptionChain, utils::OptionDataGroup},
@@ -344,9 +345,11 @@ impl BreakEvenable for BearPutSpread {
         self.break_even_points = Vec::new();
 
         self.break_even_points.push(
-            (self.long_put.option.strike_price
-                - self.get_net_cost()? / self.long_put.option.quantity)
-                .round_to(2),
+            lower_break_even(
+                self.long_put.option.strike_price,
+                self.get_net_cost()? / self.long_put.option.quantity,
+            )
+            .round_to(2),
         );
 
         Ok(())