+++ src/strategies/iron_condor.rs
@@ -19,6 +19,7 @@ use super::base::{
     BreakEvenable, Optimizable, Positionable, Strategable, StrategyBasics, StrategyType, Validable,
 };
 use super::shared::CondorStrategy;
+use crate::strategies::base::lower_break_even;
 use crate::{
     ExpirationDate, Options,
     chains::{StrategyLegs, chain::OptionChain, utils::OptionDataGroup},
@@ -471,7 +472,7 @@ impl BreakEvenable for IronCondor {
             .push((self.short_call.option.strike_price + net_credit).round_to(2));
 
         self.break_even_points
-            .push((self.short_put.option.strike_price - net_credit).round_to(2));
+            .push(lower_break_even(self.short_put.option.strike_price, net_credit).round_to(2));
 
         self.break_even_points.sort();
         Ok(())