+++ src/strategies/long_strangle.rs
@@ -19,6 +19,7 @@ use super::base::{
     BreakEvenable, Optimizable, Positionable, Strategable, StrategyBasics, StrategyType, Validable,
 };
 use super::shared::StrangleStrategy;
+use crate::strategies::base::lower_break_even;
 use crate::{
     ExpirationDate, Options,
     chains::{StrategyLegs, chain::OptionChain, utils::OptionDataGroup},
@@ -380,8 +381,11 @@ impl BreakEvenable for LongStrangle {
         let total_premium = self.get_net_cost()?;
 
         self.break_even_points.push(
-            (self.long_put.option.strike_price - (total_premium / self.long_put.option.quantity))
-                .round_to(2),
+            lower_break_even(
+                self.long_put.option.strike_price,
+                total_premium / self.long_put.option.quantity,
+            )
+            .round_to(2),
         );
 
         self.break_even_points.push(